requirements.txt
acres==0.5.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
certifi==2026.1.4
charset-normalizer==3.4.4
ci-info==0.3.0
click==8.3.1
configobj==5.0.9
configparser==7.2.0
einops==0.8.1
etelemetry==0.3.1
fastapi==0.128.0
filelock==3.20.2
fsspec==2025.12.0
h11==0.16.0
hf-xet==1.2.0
httplib2==0.31.0
huggingface-hub==0.36.0
idna==3.11
Jinja2==3.1.6
joblib==1.5.3
looseversion==1.3.0
lxml==6.0.2
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.6.1
nibabel==5.3.3
nipype==1.10.0
numpy==2.4.0
packaging==25.0
pandas==2.3.3
pathlib==1.0.1
prov==2.1.1
puremagic==1.30
pydantic==2.12.5
pydantic_core==2.41.5
pydot==4.0.1
PyMuPDF==1.26.7
pyparsing==3.3.1
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-multipart==0.0.21
pytz==2025.2
pyxnat==1.6.4
PyYAML==6.0.3
rank-bm25==0.2.2
rdflib==7.5.0
regex==2025.11.3
requests==2.32.5
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.16.3
sentence-transformers==5.2.0
setuptools==80.9.0
simplejson==3.20.2
six==1.17.0
starlette==0.50.0
sympy==1.14.0
threadpoolctl==3.6.0
tinydb==4.8.2
tokenizers==0.22.2
torch==2.9.1
tqdm==4.67.1
traits==7.1.0
transformers==4.57.3
triton==3.5.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
uvicorn==0.40.0
bcrypt
python-jose[cryptography]